% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{new_default_solver}
\alias{new_default_solver}
\title{Create a solver with solvephrust's default compounds}
\usage{
new_default_solver()
}
\value{
A new \code{solvephrust_solver} external pointer.
}
\description{
The catalog is generated from \code{data-raw/compile_data.R} and contains the
pH-dependent compounds \code{co3}, \code{po4}, \code{ocl}, \code{nh3}, \code{ch3coo}, \code{bo3}, and
\code{sio4}, and the fixed-charge compounds \code{so4}, \code{na}, \code{ca}, \code{mg}, \code{cl}, \code{mno4},
and \code{no3}.
}
